Module rtx::util::test

text-related utility functions

Functions

init_logger

lex_single_tex_formula

Simple tokenization of a single formula, without any custom preloads beyond latex and amsmath

new_test_engine

Provide a default test Core engine for simple operations

rtx_test_single

rtx_tests

rtx_tests_internal